Add files via upload
Added simple server code that can send and receive data. This server listens on port 5000 and currently can support 2 connections simultaneously. Can send and receive data  with a max size of 2048 bytes. when run the program will print out the IP address that the client will use to connect to the server.